+++ src/test/java/kitchenpos/application/MenuServiceTest.java
@@ -27,7 +27,7 @@ class MenuServiceTest {
         void 요청을_할_수_있다() {
             // given
             final MenuGroupResponse menuGroup = menuGroupService.create(new MenuGroupRequest("1인 메뉴"));
-            final ProductResponse product = productService.create(new ProductCreateRequest("짜장면", 1000));
+            final ProductResponse product = productService.create(new ProductCreateRequest("짜장면", BigDecimal.valueOf(1000)));
             final MenuCreateRequest menu = new MenuCreateRequest("짜장면", BigDecimal.valueOf(1000), menuGroup.getId(),
                 List.of(new MenuProductCreateRequest(product.getId(), 1)));
 
@@ -42,7 +42,7 @@ class MenuServiceTest {
         void 요청에서_메뉴_금액이_0원_미만인_경우_예외가_발생한다() {
             // given
             final MenuGroupResponse menuGroup = menuGroupService.create(new MenuGroupRequest("1인 메뉴"));
-            final ProductResponse product = productService.create(new ProductCreateRequest("짜장면", 1000));
+            final ProductResponse product = productService.create(new ProductCreateRequest("짜장면", BigDecimal.valueOf(1000)));
             final MenuCreateRequest menu = new MenuCreateRequest("짜장면", BigDecimal.valueOf(-1), menuGroup.getId(),
                 List.of(new MenuProductCreateRequest(product.getId(), 1)));
 
@@ -55,7 +55,7 @@ class MenuServiceTest {
         void 요청에서_메뉴_금액이_NULL_일_경우_예외가_발생한다() {
             // given
             final MenuGroupResponse menuGroup = menuGroupService.create(new MenuGroupRequest("1인 메뉴"));
-            final ProductResponse product = productService.create(new ProductCreateRequest("짜장면", 1000));
+            final ProductResponse product = productService.create(new ProductCreateRequest("짜장면", BigDecimal.valueOf(1000)));
             final MenuCreateRequest menu = new MenuCreateRequest("짜장면", null, menuGroup.getId(),
                 List.of(new MenuProductCreateRequest(product.getId(), 1)));
 
@@ -80,7 +80,7 @@ class MenuServiceTest {
         @Test
         void 요청에서_추가한_메뉴가_메뉴_그룹에_속해있지_않은경우_예외가_발생한다() {
             // given
-            final ProductResponse product = productService.create(new ProductCreateRequest("짜장면", 1000));
+            final ProductResponse product = productService.create(new ProductCreateRequest("짜장면", BigDecimal.valueOf(1000)));
             final MenuCreateRequest menu = new MenuCreateRequest("짜장면", null, null,
                 List.of(new MenuProductCreateRequest(product.getId(), 1)));
 
@@ -93,7 +93,7 @@ class MenuServiceTest {
         void 요청에서_추가한_메뉴가격의_합이_상품가격의_합보다_높을경우_예외가_발생한다() {
             // given
             final MenuGroupResponse menuGroup = menuGroupService.create(new MenuGroupRequest("1인 메뉴"));
-            final ProductResponse product = productService.create(new ProductCreateRequest("짜장면", 0));
+            final ProductResponse product = productService.create(new ProductCreateRequest("짜장면", BigDecimal.valueOf(0)));
             final MenuCreateRequest menu = new MenuCreateRequest("짜장면", BigDecimal.valueOf(1000), menuGroup.getId(),
                 List.of(new MenuProductCreateRequest(product.getId(), 1)));
 
@@ -109,7 +109,7 @@ class MenuServiceTest {
         void 요청을_할_수_있다() {
             // given
             final MenuGroupResponse menuGroup = menuGroupService.create(new MenuGroupRequest("1인 메뉴"));
-            final ProductResponse product = productService.create(new ProductCreateRequest("짜장면", 1000));
+            final ProductResponse product = productService.create(new ProductCreateRequest("짜장면", BigDecimal.valueOf(1000)));
             final MenuResponse menu = menuService.create(new MenuCreateRequest("짜장면", BigDecimal.valueOf(1000), menuGroup.getId(),
                 List.of(new MenuProductCreateRequest(product.getId(), 1))));
