+++ src/test/java/kitchenpos/application/TableServiceTest.java
@@ -66,7 +66,7 @@ class TableServiceTest {
         void 요청에서_테이블이_비어있는_상태로_변경할때_주문의_상태가_완료상태가_아니면_예외가_발생한다(final OrderStatus orderStatus) {
             // given
             final MenuGroupResponse menuGroup = menuGroupService.create(new MenuGroupRequest("1인 메뉴"));
-            final ProductResponse product = productService.create(new ProductCreateRequest("짜장면", 1000));
+            final ProductResponse product = productService.create(new ProductCreateRequest("짜장면", BigDecimal.valueOf(1000)));
             final MenuResponse menu = menuService.create(new MenuCreateRequest("짜장면", BigDecimal.valueOf(1000), menuGroup.getId(),
                 List.of(new MenuProductCreateRequest(product.getId(), 1))));
             final OrderTableResponse orderTable = tableService.create(new OrderTableCreateRequest(2, false));
